requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
pytesseract
Pillow
python-multipart
pandas
pypdfium2
passlib[bcrypt]
python-jose[cryptography]
SQLAlchemy>=2.0
rapidocr-onnxruntime
onnxruntime
paddleocr
